+++ projects/ccf-body-ui/src/lib/body-ui-layer.ts
@@ -1,7 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 import { CompositeLayer, COORDINATE_SYSTEM } from '@deck.gl/core';
+import { TextLayer } from '@deck.gl/layers';
 import { ScenegraphLayer, SimpleMeshLayer } from '@deck.gl/mesh-layers';
-import { CubeGeometry } from '@luma.gl/core';
+import { ConeGeometry, CubeGeometry, CylinderGeometry, Geometry, SphereGeometry } from '@luma.gl/core';
+import { Matrix4 } from '@math.gl/core';
 
 import { SpatialSceneNode } from './shared/spatial-scene-node';
 import { loadGLTF, loadGLTF2, registerGLTFLoaders } from './util/load-gltf';
@@ -12,6 +14,22 @@ function meshLayer(id: string, data: SpatialSceneNode[], options: { [key: string
   if (!data || data.length === 0) {
     return undefined;
   } else {
+    let mesh: Geometry;
+    switch (options.geometry) {
+      case 'sphere':
+        mesh = new SphereGeometry();
+        break;
+      case 'cone':
+        mesh = new ConeGeometry();
+        break;
+      case 'cylinder':
+        mesh = new CylinderGeometry();
+        break;
+      case 'cube':
+      default:
+        mesh = new CubeGeometry();
+        break;
+    }
     return new SimpleMeshLayer({
       ...{
         id,
@@ -21,10 +39,28 @@ function meshLayer(id: string, data: SpatialSceneNode[], options: { [key: string
         coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
         data,
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        mesh: new CubeGeometry() as any,
+        mesh: mesh as any,
         wireframe: false,
         getTransformMatrix: (d) => (d as { transformMatrix: number[][] }).transformMatrix,
         getColor: (d) => (d as { color: [number, number, number, number] }).color || [255, 255, 255, 0.9*255]
+      },
+      ...options
+    });
+  }
+}
+
+function textLayer(id: string, data: SpatialSceneNode[], options: { [key: string]: unknown }): TextLayer<unknown> | undefined {
+  if (!data || data.length === 0) {
+    return undefined;
+  } else {
+    return new TextLayer({
+      ...{
+        id,
+        pickable: true,
+        data: data.map(d => ({ ...d, position: new Matrix4(d.transformMatrix).getTranslation() })),
+        getText: (d) => (d as { text: string }).text,
+        getPosition: (d) => (d as { position: [number, number] }).position,
+        getColor: (d) => (d as { color: [number, number, number, number] }).color
       },
       ...options
     });
@@ -43,45 +79,65 @@ export class BodyUILayer extends CompositeLayer<SpatialSceneNode> {
 
   renderLayers(): unknown[] {
     const state = this.state as { data: SpatialSceneNode[]; zoomOpacity: number; doCollisions: boolean };
-    const cubes = state.data.filter(d => !d.scenegraph && !d.wireframe && d.unpickable);
-    const pickableCubes = state.data.filter(d => !d.scenegraph && !d.wireframe && !d.unpickable);
-    const wireframes = state.data.filter(d => !d.scenegraph && d.wireframe);
-    const models = state.data.filter(d => !!d.scenegraph);
+    const geometries: Record<string, SpatialSceneNode[]> = {
+      'sphere': [], 'cone': [], 'cylinder': [], 'cube': [], 'text': [], 'wireframe': [], 'scenegraph': []
+    };
 
-    if (state.doCollisions) {
-      doCollisions(state.data);
+    for (const node of state.data) {
+      const geometry = node.geometry ?? 'cube';
+      if (node.scenegraph) {
+        geometries.scenegraph.push(node);
+      } else if (geometries[geometry] !== undefined) {
+        geometries[geometry].push(node);
+      }
     }
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const url2gltf: { [url: string]: Promise<any> } = {};
-    for (const m of models) {
+    for (const m of geometries.scenegraph) {
       if (m.scenegraph && m.scenegraphNode && !Object.prototype.hasOwnProperty.call(url2gltf, m.scenegraph)) {
         url2gltf[m.scenegraph] = loadGLTF({ scenegraph: m.scenegraph } as SpatialSceneNode, BodyUILayer.gltfCache);
       }
     }
 
-    return [
-      meshLayer('cubes', cubes, { wireframe: false, pickable: false }),
-      meshLayer('pickableCubes', pickableCubes, { wireframe: false, pickable: true }),
-      meshLayer('wireframes', wireframes, { wireframe: true, pickable: false }),
-      ...models.map((model) =>
-        new ScenegraphLayer({
-          id: 'models-' + model['@id'],
-          opacity: model.zoomBasedOpacity ? state.zoomOpacity : (model.opacity !== undefined ? model.opacity : 1.0),
-          pickable: !model.unpickable,
-          coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-          data: [model],
-          scenegraph: model.scenegraphNode ?
-            loadGLTF2(model.scenegraphNode, url2gltf[model.scenegraph as string]) :
-            model.scenegraph as unknown as URL,
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          _lighting: model._lighting, // 'pbr' | undefined
-          getTransformMatrix: model.transformMatrix as unknown as number[][],
-          getColor: model.color ?? [0, 255, 0, 0.5*255],
-          parameters: { depthMask: !model.zoomBasedOpacity && (model.opacity === undefined || model.opacity === 1) }
-        })
-      )
-    ].filter(l => !!l);
+    const layers: unknown[] = [];
+    for (const [geometry, nodes] of Object.entries(geometries)) {
+      if (geometry === 'scenegraph') {
+        for (const model of nodes) {
+          layers.push(
+            new ScenegraphLayer({
+              id: 'models-' + model['@id'],
+              opacity: model.zoomBasedOpacity ? state.zoomOpacity : (model.opacity !== undefined ? model.opacity : 1.0),
+              pickable: !model.unpickable,
+              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
+              data: [model],
+              scenegraph: model.scenegraphNode ?
+                loadGLTF2(model.scenegraphNode, url2gltf[model.scenegraph as string]) :
+                model.scenegraph as unknown as URL,
+              // eslint-disable-next-line @typescript-eslint/naming-convention
+              _lighting: model._lighting, // 'pbr' | undefined
+              getTransformMatrix: model.transformMatrix as unknown as number[][],
+              getColor: model.color ?? [0, 255, 0, 0.5*255],
+              parameters: { depthMask: !model.zoomBasedOpacity && (model.opacity === undefined || model.opacity === 1) }
+            })
+          );
+        }
+      } else if (geometry === 'text') {
+        layers.push(textLayer('text', nodes.filter(n => n.unpickable), { pickable: false }));
+        layers.push(textLayer('textPickable', nodes.filter(n => !n.unpickable), { pickable: true }));
+      } else if (geometry === 'wireframe') {
+        layers.push(meshLayer(geometry, nodes, { wireframe: true, pickable: false, geometry }));
+      } else {
+        layers.push(meshLayer(geometry, nodes.filter(n => n.unpickable), { wireframe: false, pickable: false, geometry }));
+        layers.push(meshLayer(`${geometry}Pickable`, nodes.filter(n => !n.unpickable), { wireframe: false, pickable: true, geometry }));
+      }
+    }
+
+    if (state.doCollisions) {
+      doCollisions(state.data);
+    }
+
+    return layers.filter(l => !!l);
   }
 
   getPickingInfo(