+++ projects/ccf-eui/src/app/shared/components/spatial-search-config/spatial-search-config.component.ts
@@ -1,0 +1,29 @@
+import { ChangeDetectionStrategy, Component, EventEmitter, HostBinding, Input, Output } from '@angular/core';
+import { OrganInfo } from 'ccf-shared';
+
+
+export type Sex = 'male' | 'female';
+
+@Component({
+  selector: 'ccf-spatial-search-config',
+  templateUrl: './spatial-search-config.component.html',
+  styleUrls: ['./spatial-search-config.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class SpatialSearchConfigComponent {
+  @HostBinding('class') readonly className = 'ccf-spatial-search-config';
+
+  @Input() organs: OrganInfo[];
+
+  @Input() selectedOrgan?: OrganInfo;
+
+  @Input() sex: Sex;
+
+  @Output() readonly updateSex = new EventEmitter<Sex>();
+
+  @Output() readonly updateOrgan = new EventEmitter<OrganInfo>();
+
+  @Output() readonly buttonClicked = new EventEmitter();
+
+  @Output() readonly closeDialog = new EventEmitter();
+}