+++ projects/ccf-eui/src/app/shared/components/spatial-search-ui/spatial-search-ui.module.ts
@@ -1,0 +1,15 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { TissueBlockListModule } from '../tissue-block-list/tissue-block-list.module';
+
+import { SpatialSearchUiComponent } from './spatial-search-ui.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    TissueBlockListModule
+  ],
+  declarations: [SpatialSearchUiComponent],
+  exports: [SpatialSearchUiComponent]
+})
+export class SpatialSearchUiModule { }